docs clean up: remove requirements.txt (no longer needed), and manually added module.rst files (using autosummary recursive, although it does not work so well)

+++ particles/distributions.py
@@ -228,7 +228,7 @@ class ProbDist:
 
     """
     dim = 1  # distributions are univariate by default
-    dtype = np.float64  # distributions are continuous by default
+    dtype = float  # distributions are continuous by default
 
     def shape(self, size):
         if size is None:
